image_webp

Module alpha_blending

Optimized alpha blending routines based on libwebp

https://github.com/webmproject/libwebp/blob/e4f7a9f0c7c9fbfae1568bc7fa5c94b989b50872/src/demux/anim_decode.c#L215-L267

Functions

blend_channel_nonpremult 🔒

Blend a single channel of src over dst, given their alpha channel values. src and dst are assumed to be NOT pre-multiplied by alpha.

blend_pixel_nonpremult 🔒

Blend src over dst assuming they are NOT pre-multiplied by alpha.

channel_shift 🔒

div_by_255 🔒

Divides by 255, rounding to nearest (as opposed to down, like regular integer division does). TODO: cannot output 256, so the output is effecitively u8. Plumb that through the code.

do_alpha_blending 🔒