image_webp

Module loop_filter

Does loop filtering on webp lossy images

Functions

c 🔒

common_adjust_horizontal 🔒

Used in both the simple and normal filters described in 15.2 and 15.3

common_adjust_vertical 🔒

Used in both the simple and normal filters described in 15.2 and 15.3

diff 🔒

high_edge_variance_horizontal 🔒

high_edge_variance_vertical 🔒

macroblock_filter_horizontal 🔒

Part of the normal filter described in 15.3 in the specification

macroblock_filter_vertical 🔒

Part of the normal filter described in 15.3 in the specification

s2u 🔒

should_filter_horizontal 🔒

should_filter_vertical 🔒

simple_segment_horizontal 🔒

Part of the simple filter described in 15.2 in the specification

simple_segment_vertical 🔒

Part of the simple filter described in 15.2 in the specification

simple_threshold_horizontal 🔒

simple_threshold_vertical 🔒

subblock_filter_horizontal 🔒

Part of the normal filter described in 15.3 in the specification

subblock_filter_vertical 🔒

Part of the normal filter described in 15.3 in the specification

u2s 🔒