Crate image_webp

Decoding and Encoding of WebP Images

Modules

vp8

An implementation of the VP8 Video Codec

Structs

EncoderParams

Allows fine-tuning some encoder parameters.

WebPDecodeOptions

WebP decoder configuration options

WebPDecoder

WebP image format decoder.

WebPEncoder

WebP Encoder.

Enums

ColorType

Color type of the image.

DecodingError

Errors that can occur when attempting to decode a WebP image

EncodingError

Error that can occur during encoding.

LoopCount

Number of times that an animation loops.

UpsamplingMethod

Methods for upsampling the chroma values in lossy decoding