image_webp

Module decoder

Structs

AnimationState 🔒

WebPDecodeOptions

WebP decoder configuration options

WebPDecoder

WebP image format decoder.

Enums

DecodingError

Errors that can occur when attempting to decode a WebP image

ImageKind 🔒

LoopCount

Number of times that an animation loops.

UpsamplingMethod

Methods for upsampling the chroma values in lossy decoding

WebPRiffChunk 🔒

All possible RIFF chunks in a WebP image file

Functions

range_reader 🔒

read_chunk_header 🔒

read_fourcc 🔒